Add partialExportUtils and useEntitesToExport files
